Added material-icons, fixed project path
