+++ react-chat/src/components/CreateChatButton.js
@@ -1,10 +1,11 @@
 import React from "react";
 import {CreateButton} from "../styles/PageChatListStyles";
+import AddCircleIcon from '@mui/icons-material/AddCircle';
 
 export const CreateChatButton = () => {
     return (
         <CreateButton>
-            <span className="material-symbols-outlined">add_circle</span>
+            <AddCircleIcon/>
         </CreateButton>
     )
 }
