+++ react-chat/src/components/PageChatListHeader.js
@@ -1,18 +1,16 @@
 import React from "react";
 import {Button, Header, PageTitle} from "../styles/PageChatListStyles";
+import MenuIcon from '@mui/icons-material/Menu';
+import SearchIcon from '@mui/icons-material/Search';
 
 export const PageChatListHeader = () => {
     return (
         <Header>
-            <Button>
-                <span className="material-icons">menu</span>
-            </Button>
+            <Button><MenuIcon/></Button>
             <PageTitle>
                 <span>Messenger</span>
             </PageTitle>
-            <Button>
-                <span className="material-icons">search</span>
-            </Button>
+            <Button><SearchIcon/></Button>
         </Header>
     )
 }
